Add zoom choices

+++ src/classes/Dimensoes.java
@@ -1,8 +1,8 @@
 package classes;
 
 public class Dimensoes {
-	public static short LARGURA;
-	public static short ALTURA;
-	public static short ALTURA_CAPA;
-	public static short ZOOM = 2;
+	public static int LARGURA;
+	public static int ALTURA;
+	public static int ALTURA_CAPA;
+	public static int ZOOM = 2;
 }
